# MM-25710: Use an efficient cache serialization algorithm