--- a/services/cache2/lru_test.go
+++ b/services/cache2/lru_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/mattermost/mattermost-server/v5/model"
-	"github.com/mattermost/mattermost-server/v5/services/cache/lru"
 	"github.com/stretchr/testify/require"
 )