--- a/services/cache2/lru_test.go
+++ b/services/cache2/lru_test.go
@@ -9,6 +9,8 @@ import (
 	"time"
 
 	"github.com/mattermost/mattermost-server/v5/model"
+	"github.com/mattermost/mattermost-server/v5/services/cache/lru"
+
 	"github.com/stretchr/testify/require"
 )