@streamer45 2020-06-17
> @streamer45 - Sorry, could you clarify? Do you mean the benchmark results or the benchmark code? The results are there in the first commit, and the benchmark code is already there in the codebase.

I was talking about the actual code. I've implicitly thought it was written for this PR.
Found it now, thanks :+1: