@agnivade 2020-06-18
@streamer45 - I gave cbor a spin. Here are the results compared against the PR's algorithm:

I could not use a sync.Pool with encoders/decoders because the API does not expose a reset functionality. Instead the generic Marshal/Unmarshal functions use some cached state pool internally, so I let it use that.

```
name               old time/op    new time/op    delta
LRU/simple=new-8     3.68µs ± 3%    3.54µs ± 2%   -3.66%  (p=0.000 n=10+8)
LRU/complex=new-8    8.67µs ± 3%    9.76µs ± 3%  +12.61%  (p=0.000 n=10+10)
LRU/User=new-8       20.9µs ± 1%    24.4µs ± 2%  +16.86%  (p=0.000 n=10+9)
LRU/Post=new-8       21.2µs ± 2%    15.9µs ± 2%  -25.07%  (p=0.000 n=10+10)
LRU/Status=new-8     5.05µs ± 3%    4.63µs ± 4%   -8.36%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
LRU/simple=new-8     1.60kB ± 0%    1.46kB ± 0%   -9.20%  (p=0.000 n=10+10)
LRU/complex=new-8    4.58kB ± 0%    3.75kB ± 0%  -18.09%  (p=0.000 n=8+8)
LRU/User=new-8       10.3kB ± 0%     8.7kB ± 0%  -15.00%  (p=0.000 n=9+10)
LRU/Post=new-8       5.95kB ± 0%    4.26kB ± 0%  -28.47%  (p=0.000 n=10+10)
LRU/Status=new-8     1.86kB ± 0%    1.66kB ± 0%  -10.97%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
LRU/simple=new-8       20.0 ± 0%      18.0 ± 0%  -10.00%  (p=0.000 n=10+10)
LRU/complex=new-8      46.0 ± 0%      67.0 ± 0%  +45.65%  (p=0.000 n=10+10)
LRU/User=new-8          106 ± 0%       169 ± 0%  +59.43%  (p=0.000 n=10+10)
LRU/Post=new-8         89.0 ± 0%      85.0 ± 0%   -4.49%  (p=0.000 n=10+10)
LRU/Status=new-8       23.0 ± 0%      21.0 ± 0%   -8.70%  (p=0.000 n=10+10)
```

As we can see, it is faster for smaller structs, but slower for larger ones. But even apart from that, the encoding itself is slightly problematic. If we store a `map[string]interface{}` with an int value, that gets stored and then decoded to a hex value. I believe this is because of cbor's algorithm which is geared towards small data sizes.

This maybe problematic in some of our code where we use a `map[string]interface{}` as props, and it might fail in mysterious ways.