@streamer45 2020-06-18
> 
> As we can see, it is faster for smaller structs, but slower for larger ones. But even apart from that, the encoding itself is slightly problematic. If we store a `map[string]interface{}` with an int value, that gets stored and then decoded to a hex value. I believe this is because of cbor's algorithm which is geared towards small data sizes.
> 
> This maybe problematic in some of our code where we use a `map[string]interface{}` as props, and it might fail in mysterious ways.

That's interesting. From the little I've spent on it my understanding was that a `map[string]interface{}` with `int` should return either a `int64` or `uint64`. (e.g. https://play.golang.org/p/N0NcDRncvxU).

As far as pure speed there might also be some encoding options and other optimizations that can be used (`keyasint` and `toarray`) to speed it up further.

That said, I am sure we can revisit this later if it makes sense to do so.

Thanks for taking the time to look into this :+1: