@agnivade 2020-06-18
Ooh right, it was an uint. Sorry I didn't consider checking its type.

Yeah, I think the next step would be to dynamically switch to msgp for hot structs. That avoids reflection altogether and will be much faster, while still keeping the `int` compatibility. We just need to rearrange the code a bit for that. What needs to be seen is if the code changes are worth it or not.